+++ doc/user/changelog.rst
@@ -4,6 +4,8 @@ Kenozooid 0.16.0
 ----------------
 - decompression dive plan summary extended with O2 partial pressure and
   maximum operating depth information
+- reordered decompression dive plan profiles to have lost gas emergency
+  scenario after planned dive profile
 - calculator functions `ppO2` and `ppN2` renamed to `pp_o2` and `pp_n2`
 
 Kenozooid 0.15.1